requirements.txt
requests
feedparser
colorama
pyyaml
pyrate-limiter
listparser
lxml
html2text